Added support for global value offsets
This commit depends on this PR landing on cranelift:
https://github.com/CraneStation/cranelift/pull/612

+++ src/webassembly/import_object.rs
@@ -112,7 +112,7 @@ where
 #[derive(PartialEq, Debug)]
 pub enum ImportValue {
     Func(*const u8),
-    Global(u8),
+    Global(i64),
     Table(Vec<usize>),
     Memory(LinearMemory),
 }
